# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Transport Management System.",
        "title": "Transport Management System",
        "contact": {}
    },
    "paths": {
...
        "/api/main/v1/tms-role/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "api for create tms role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tms - Role"
                ],
                "summary": "Create Tms Role",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CreateTmsRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.CreateTmsRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.JSONResponse"
                        }
                    }
                }
            }
        },
        "/api/main/v1/tms-role/{tmsRoleId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "api for Get Tms Role  by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tms - Role"
                ],
                "summary": "Get Tms Role  By Id",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tms Role Id",
                        "name": "tmsRoleId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.GetDetailTmsRolesbyIdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.JSONResponse"
                        }
                    }
                }
            }
        },
        "/api/main/v1/tms-role/{tmsRoleId}/delete": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "api for delete Tms Role by Id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tms - Role"
                ],
                "summary": "Delete Tms Role",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tms Role Id",
                        "name": "tmsRoleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "delete data for tms role",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.DeleteTmsRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.DeleteTmsRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.DeleteTmsRoleResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.DeleteTmsRoleResponse"
                        }
                    }
                }
            }
        },
        "/api/main/v1/tms-role/{tmsRoleId}/update": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "api for update tms role by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tms - Role"
                ],
                "summary": "Update Tms Role",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tms Role Id",
                        "name": "tmsRoleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data for tms role",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.UpdateTmsRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.DetailTmsRole"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.JSONResponse"
                        }
                    }
                }
            }
        },
        "/api/main/v1/tms-roles": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "api for Get list tms role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tms - Role"
                ],
                "summary": "Get List Tms Role",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.GetTmsRolesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.GetTmsRolesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/helper.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/helper.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bluebird_tech_kirim_protobuf_auth_v1.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "bluebird_tech_kirim_protobuf_auth_v1.PaginationWithTotalRows": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
        "bluebird_tech_kirim_protobuf_auth_v1.ResponseStatus": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "latency": {
                    "type": "integer"
                },
                "response_time": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "helper.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "pagination": {},
                "response_status": {}
            }
        },
...
        "v1.CreateTmsRoleRequest": {
            "type": "object",
            "properties": {
                "actor_user_id": {
                    "type": "string"
                },
                "app_domain": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "permission_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
        "v1.CreateTmsRoleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.DataIdResponse"
                },
                "response_status": {
                    "$ref": "#/definitions/bluebird_tech_kirim_protobuf_auth_v1.ResponseStatus"
                }
            }
        },
...
        "v1.DataIdResponse": {
            "type": "object",
            "properties": {
                "data_id": {
                    "type": "string"
                }
            }
        },
...
        "v1.DeleteTmsRoleRequest": {
            "type": "object",
            "properties": {
                "actor_user_id": {
                    "type": "string"
                },
                "role_id": {
                    "type": "string"
                },
                "row_version": {
                    "type": "string"
                }
            }
        },
        "v1.DeleteTmsRoleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.DataIdResponse"
                },
                "response_status": {
                    "$ref": "#/definitions/bluebird_tech_kirim_protobuf_auth_v1.ResponseStatus"
                }
            }
        },
...
        "v1.DetailTmsRole": {
            "type": "object",
            "properties": {
                "app_domain": {
                    "type": "string"
                },
                "created_time_utc": {
                    "type": "string"
                },
                "created_user_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "permission": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.RolePermission"
                    }
                },
                "role_id": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                },
                "row_version": {
                    "type": "string"
                },
                "updated_time_utc": {
                    "type": "string"
                },
                "updated_user_id": {
                    "type": "string"
                }
            }
        },
...
        "v1.GetDetailTmsRolesbyIdResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.DetailTmsRole"
                },
                "response_status": {
                    "$ref": "#/definitions/bluebird_tech_kirim_protobuf_auth_v1.ResponseStatus"
                }
            }
        },
...
        "v1.GetTmsRolesFilter": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
        "v1.GetTmsRolesRequest": {
            "type": "object",
            "properties": {
                "filter": {
                    "$ref": "#/definitions/v1.GetTmsRolesFilter"
                },
                "pagination": {
                    "$ref": "#/definitions/bluebird_tech_kirim_protobuf_auth_v1.Pagination"
                },
                "sort": {
                    "$ref": "#/definitions/v1.GetTmsRolesSort"
                }
            }
        },
        "v1.GetTmsRolesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.TmsRole"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/bluebird_tech_kirim_protobuf_auth_v1.PaginationWithTotalRows"
                },
                "response_status": {
                    "$ref": "#/definitions/bluebird_tech_kirim_protobuf_auth_v1.ResponseStatus"
                }
            }
        },
        "v1.GetTmsRolesSort": {
            "type": "object",
            "properties": {
                "created_time_utc": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                },
                "updated_time_utc": {
                    "type": "string"
                }
            }
        },
...
        "v1.RolePermission": {
            "type": "object",
            "properties": {
                "app_domain": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "permission_group": {
                    "type": "string"
                },
                "permission_id": {
                    "type": "string"
                },
                "permission_name": {
                    "type": "string"
                },
                "permission_sub_group": {
                    "type": "string"
                }
            }
        },
...
        "v1.TmsRole": {
            "type": "object",
            "properties": {
                "app_domain": {
                    "type": "string"
                },
                "created_time_utc": {
                    "type": "string"
                },
                "created_user_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "role_id": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                },
                "row_version": {
                    "type": "string"
                },
                "updated_time_utc": {
                    "type": "string"
                },
                "updated_user_id": {
                    "type": "string"
                }
            }
        },
...
        "v1.UpdateTmsRoleRequest": {
            "type": "object",
            "properties": {
                "actor_user_id": {
                    "type": "string"
                },
                "app_domain": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "permission_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "role_id": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                },
                "row_version": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AppIdAuth": {
            "type": "apiKey",
            "name": "app-id",
            "in": "header"
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ]
}